refactor:create package analysis-utils
