+++ packages/analysis-utils/README.md
@@ -1,0 +1,4 @@
+# Analysis Utils
+
+## Description
+A package for common utilities for analysing websites.
