+++ packages/analysis-utils/src/browserManagement/parseNetworkDataToCookieData.ts
