+++ packages/analysis-utils/src/procedures/analyzeCookieUrlsInBatches.ts
@@ -19,48 +19,59 @@
  */
 // @ts-ignore Package does not support typescript.
 import Spinnies from 'spinnies';
-// @ts-ignore Package does not support typescript.
-import Wapplalyzer from 'wappalyzer';
 
 /**
  * Internal dependencies.
  */
-import { TechnologyDetailList } from '../types';
+import { CookieDatabase } from '../types';
+import { Cookie } from '../browserManagement/types';
+import analyzeCookiesUrls from './analyzeCookieUrls';
 
-export const analyzeTechnologiesUrlsInBatches = async (
-  urls: Array<string>,
+const analyzeCookiesUrlsInBatches = async (
+  urls: string[],
+  isHeadless: boolean,
+  delayTime: number,
+  cookieDictionary: CookieDatabase,
   batchSize = 3
-): Promise<TechnologyDetailList[]> => {
+) => {
   const spinnies = new Spinnies();
-
-  let report: TechnologyDetailList[] = [];
-  const wappalyzer = new Wapplalyzer();
+  let report: {
+    pageUrl: string;
+    cookieData: {
+      [frameUrl: string]: {
+        cookiesCount: number;
+        frameCookies: {
+          [key: string]: Cookie;
+        };
+      };
+    };
+  }[] = [];
 
   for (let i = 0; i < urls.length; i += batchSize) {
     const start = i;
     const end = Math.min(urls.length - 1, i + batchSize - 1);
-    await wappalyzer.init();
 
-    spinnies.add(`spinner-${i}`, {
-      text: `Analysing technologies in urls ${start + 1} - ${end + 1} `,
+    spinnies.add(`cookies-spinner-${i}`, {
+      text: `Analysing cookies in urls ${start + 1} - ${end + 1} `,
     });
 
     const urlsWindow = urls.slice(start, end + 1);
 
-    const technologyAnalysis = await Promise.all(
-      urlsWindow.map(async (url) => {
-        const { technologies } = await (await wappalyzer.open(url)).analyze();
-        return technologies;
-      })
+    const cookieAnalysis = await analyzeCookiesUrls(
+      urlsWindow,
+      isHeadless,
+      delayTime,
+      cookieDictionary
     );
 
-    report = [...report, ...technologyAnalysis];
+    report = [...report, ...cookieAnalysis];
 
-    spinnies.succeed(`spinner-${i}`, {
-      text: `Done technology in urls ${start + 1} - ${end + 1} `,
+    spinnies.succeed(`cookies-spinner-${i}`, {
+      text: `Done analysing cookies in urls ${start + 1} - ${end + 1} `,
     });
-    await wappalyzer.destroy();
   }
 
   return report;
 };
+
+export default analyzeCookiesUrlsInBatches;
