+++ packages/analysis-utils/src/utils/index.ts
@@ -14,24 +14,5 @@
  * limitations under the License.
  */
 
-import http from 'http';
-
-/**
- * @param port number Port number to test
- * @returns Promise which will resolve in a boolean value
- */
-function checkPortInUse(port: number): Promise<boolean> {
-  return new Promise((resolve) => {
-    const server = http
-      .createServer()
-      .listen(port, () => {
-        server.close();
-        resolve(false);
-      })
-      .on('error', () => {
-        resolve(true);
-      });
-  });
-}
-
-export default checkPortInUse;
+export { default as delay } from './delay';
+export { default as findAnalyticsMatch } from './findAnalyticsMatch';
