+++ packages/cli-dashboard/tsconfig.json
@@ -1,7 +1,8 @@
 {
   "extends": "../../tsconfig.shared.json",
   "compilerOptions": {
-    "rootDir": "src"
+    "rootDir": "src",
+    "types": ["node"]
   },
   "include": [
     "src",
