+++ packages/cli/package.json
@@ -20,24 +20,15 @@
   "homepage": "https://github.com/GoogleChromeLabs/ps-analysis-tool",
   "dependencies": {
     "@ps-analysis-tool/common": "*",
-    "cli-color": "^2.0.3",
-    "cli-table": "^0.3.11",
+    "@ps-analysis-tool/analysis-utils": "*",
     "commander": "^10.0.1",
     "fs-extra": "^11.1.1",
-    "ora": "^7.0.1",
-    "promptly": "^3.2.0",
-    "puppeteer": "^21.0.3",
-    "simple-cookie": "^1.0.15",
-    "sitemapper": "^3.1.8",
     "spinnies": "^0.5.1",
-    "tldts": "^6.0.14",
-    "wappalyzer": "^6.10.66"
+    "promptly": "^3.2.0"
   },
   "devDependencies": {
-    "@types/cli-color": "^2.0.2",
     "@types/fs-extra": "^11.0.1",
     "@types/node": "^20.2.3",
-    "@types/promptly": "^3.0.2",
     "tsc-watch": "^6.0.4",
     "typescript": "^5.0.4"
   }
