+++ packages/cli/src/utils/index.ts
@@ -14,3 +14,6 @@
  * limitations under the License.
  */
 export { default as delay } from './delay';
+export { default as fetchCookieDictionary } from './fetchCookieDictionary';
+export { default as checkPortInUse } from './checkPortInUse';
+export { default as Utility } from './utility';
