+++ packages/cli/tsconfig.json
@@ -9,5 +9,6 @@
     "esModuleInterop": true,
     "moduleResolution": "node"
   },
-  "references": [{ "path": "../common" }]
+  "exclude": ["dist/**, dist-types/**"],
+  "references": [{ "path": "../common" }, { "path": "../analysis-utils" }]
 }
